requirements.txt
PyPDF2>=3.0.0
langchain>=0.1.0
langchain-core>=0.3.49
pinecone-client>=3.0.0
requests>=2.31.0
tqdm>=4.66.0
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
einops>=0.7.0
